Added new makePotionEffect() and bukkitColor() functions
makePotionEffect() can be used for easily making PotionEffect objects that can be applied to any mob or player with lots of options to customize the effect to suit your needs. bukkitColor() is used to make a Color object (org.bukkit.Color) from given RGB values.. some methods in the spigot API take this object as an argument, including PotionEffect.