Refactoring - new approach
Since the legacy and latest Bukkit API share most of the code base, it's
exhausing to separate every single implementation, and in fact, doing so
cause lots of boilerplate codes instead of nice, coherent codes. As a
first step, the codes from bukkit-legacy are moved to its parent,
latest, so that both bukkit-legacy and bukkit-latest can have chance to
share the code base which otherwise will be almost identical boilerplace
codes unnecessarily.

Then, there are few classes (such as CommonFunctions) that has to be
implemented separately; those classes are implemented separately in each
project.